2022 d10 - combine p1 & p2 for perf
This was already very fast, but there's always more perf to squeeze out.

I don't think I can even reliably say if or how much faster this is
without instrumenting the code itself, which I don't feel like doing
right now. My lazy instrumentation of my perf on these problems is
usually just using `time` on the command line, but granularity there is
only down to the ms and this was already at about 1ms. This certainly
*should* be faster, even if we're talking about ns at this point, so I'm
keeping it.