2022 d11 p2
The "keep your worry levels manageable" bit was tricky. I realized we
needed to take the modulo of the item worry level to something, but
wasn't sure how to figure out the "something" so that the remainder
would still be congruent for all the other monkeys. I.e. if monkey 0
tests mod 3 and another monkey tests mod 7, you need to mod by something
such that future mod operations on 3 & 7 would have the same result.

All the numbers the monkeys used for their modulo tests are prime, so
there's probably some theorem I'm not thinking of that is applicable
here. But I was pretty sure if I just used the lcm of all the monkey
modulo test numbers that would also work. Rust stdlib doesn't have
an `lcm` that I know of & I didn't want to write one, so I just
multiplied all the numbers together, which is at least *a* CM, if not
*the* LCM.

Oh, wait: because the numbers are all prime just multiplying them
together *is* the LCM. Every number has a unique prime factorization, so
therefore the LCM of a bunch of primes is just their product.

So yes, calculating the LCM and then modding the numbers by that works.
It did require me to use u128s (u64s still overflowed), so I bet there
is another more clever solution that keeps things even smaller.

Reminding myself of the Chinese Remainder Theorem, I think that's
applicable here & could be used to determine the modulus of a number
with all the prime divisors here without actually carrying around the
large LCM & dealing with larger numbers. But thankfully rust does have
native 128 bit types and the somewhat naive LCM approach worked, so I'm
calling it good.