2022 d12
First path finding puzzle of the year. I think p2 is a different
variation than we usually get, so that was fun.

This is my slowest run-time solution so far this year by a wide margin
at 300ms. Which is still plenty fast, but I think every other day this
year has been < 50ms. I may come back later and rework this for
performance. I think the most performant approach would be to
effectively do the search in reverse: "start" at the goal and search to
all the possible 'a' spots, with the appropriate reversal to the "can
step" logic. Doing it that way would allow effectively only running
djikstra once: just return the map of distances at the end rather than
early-returning when we find a goal.