2022 d12 perf improvement
As described in previous commit, it's much faster to run djikstra once
to build all paths & then just find the relevant ones.

As initially written for p1 that wasn't immediately feasible, because
for p1 I did the logical thing of finding a path from the start to the
goal, but to build all the paths relevant to p1 & p2 in one pass you
want to go backwards: from goal to every square reachable from there
(with the definition of "reachable" the reverse of problem description).
Or in other words you back-track from the goal to every position that
could walk to the goal.

Did that refactor, and this dropped from 300ms to single-digit ms.

Because we are no longer early-returning at a goal but building the
entire possible set of paths, we also no longer need a heap for the
queue: a simple deque is sufficient. Makes no measurable difference in
run time, but is a bit less code.