2022 d14 p1
I was initially going to just use a `HashMap` for grid storage, since the
input makes it difficult to know what bounds to use for vecs up-front
for a `Vec<Vec<_>>` grid. But I've had this thought before about
dynamically growing and adjusting nested vecs as the grid grows, so I
thought I'd try that out. Pretty sure this is the first time I've
successfully used lifetimes, too.