2022 d15 final
This was one of those days where actually trying to model the problem
space as data structures was harmful. Much more efficient to just count
ranges with math.

This has the slowest runtime this year for me so far by a large margin:
right around 300ms. This was a challenging one today, so I'm not going
to spend much more time on optimization.